moved form components to pages
