--- components/form-pages/about-the-waste/quantity-of-waste/index.tsx
@@ -1,1 +1,0 @@
-export const QuantityOfWasteForm = () => {};
