--- components/form-pages/about-the-waste/waste-codes-form/styled-components.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const CaptionWrapper = styled.div`
-    margin-bottom: 1rem;
-`;
