+++ components/form-pages/styled-components.ts
@@ -3,3 +3,7 @@ import styled from 'styled-components';
 export const ButtonWrapper = styled.div`
     padding-top: 1rem;
 `;
+
+export const CaptionWrapper = styled.div`
+    margin-bottom: 1rem;
+`;
