+++ pages/submit-an-export/about-the-waste/collection-date/index.tsx
@@ -1,0 +1,5 @@
+import { NextPage } from 'next';
+
+const CollectionDate: NextPage = () => {
+  return <></>;
+};
