+++ pages/submit-an-export/about-the-waste/waste-codes-descriptions/ewc-code-summary/index.tsx
@@ -1,0 +1,9 @@
+import { NextPage } from 'next';
+
+const EWCCodeSummary: NextPage = () => {
+    return(
+        <></>
+    );
+}
+
+export default EWCCodeSummary;
