--- pages/submit-an-export/about-the-waste/waste-codes-descriptions/ewc-code/index.tsx
@@ -1,8 +1,0 @@
-import { NextPage } from 'next';
-import { EWCCodeForm } from '../../../../../components/form-pages/about-the-waste/waste-codes-form/ewc-code-form/index'
-
-const EWCCode: NextPage = () => {
-  return <EWCCodeForm />
-};
-
-export default EWCCode;
