+++ pages/submit-an-export/index.tsx
@@ -1,11 +1,11 @@
 import { NextPage } from 'next';
-import { UniqueRefForm } from '../../components/form-pages/unique-ref-form';
+// import { UniqueRefForm } from '../../components/form-pages/unique-ref-form';
 import { Main } from 'govuk-react';
 
 const SubmitExport: NextPage = () => {
   return (
     <Main>
-      <UniqueRefForm />
+      {/* <UniqueRefForm /> */}
     </Main>
   );
 };
